refactor: move form to another url
